+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.2 (June 2, 2017)
+
+* Fix to Sorting
+* Added Multicolumn Sorting
+* Fix dynamic adding rows with update to interface
+* Ajax with multicolumn sorting
+
 ## 1.1.1 (June 1, 2017)
 
 * Added more Events